+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 **Fixed**
 - Operator was failing to remove finalizers from ApiDefinition that was already deleted in Dashboard [#469](https://github.com/TykTechnologies/tyk-operator/issues/469)
 - Fixed the problem of linking existing security policies while migration [#204](https://github.com/TykTechnologies/tyk-operator/issues/204)
+- Fix Security Policy tests
 
 
 ## [v0.12.0](https://github.com/TykTechnologies/tyk-operator/tree/v0.12.0)